2_DocIntelligence-AzFunction/requirements.txt
azure-ai-formrecognizer
azure-storage-blob